requirements.txt
fastapi
uvicorn
pillow
numpy
requests
langchain
langchain-google-genai
python-multipart
pydantic